+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 * 2.1.7:
     * Fix: Do not add thousand separator to paypal amounts.
+    * Fix: Improve product and order search mechanism.
 * 2.1.6.1 - 2017.09.21:
     * Fix: Added BN code to paypal plugin.
 * 2.1.6 - 2017.09.20: